requirements.txt
Django~=2.0.0
channels~=2.0,>=2.0.2
channels_redis~=2.0
django-registration-redux
django-autocomplete-light
django-taggit

nibabel
SimpleITK
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-pandas==0.5.1
djangorestframework==3.8.2
django-oauth-toolkit

docutils==0.14
et-xmlfile==1.0.1
future==0.16.0
h5py==2.7.1
jmespath==0.9.3
opencv-python==3.4.0.12
pandas==0.22.0
Pillow==5.1.0
python-dateutil==2.7.2
pytz==2018.4
rest-pandas==1.0.0
s3transfer==0.1.13
six==1.11.0
wincertstore==0.2
boto3==1.7.7
botocore==1.10.7
certifi==2018.4.16

beautifulsoup4==4.6.0
lxml==4.1.0

javabridge==1.0.14
python-bioformats==1.3.2

scipy
